+++ src/main/java/ru/practicum/shareit/item/service/ItemService.java
@@ -5,13 +5,13 @@ import ru.practicum.shareit.item.dto.ItemDto;
 import java.util.List;
 
 public interface ItemService {
-    ItemDto addItem(int userId, ItemDto itemDto);
+    ItemDto addItem(Long userId, ItemDto itemDto);
 
-    ItemDto updateItem(int ownerId, int itemId, ItemDto itemDto);
+    ItemDto updateItem(Long ownerId, Long itemId, ItemDto itemDto);
 
-    ItemDto getItemById(int itemId);
+    ItemDto getItemById(Long itemId);
 
-    List<ItemDto> getAllOwnersItems(int ownerId);
+    List<ItemDto> getAllOwnersItems(Long ownerId);
 
     List<ItemDto> searchItems(String text);
 }
