+++ src/main/java/ru/practicum/shareit/user/service/UserService.java
@@ -7,11 +7,11 @@ import java.util.List;
 public interface UserService {
     User addUser(User user);
 
-    User updateUser(int userId, User user);
+    User updateUser(Long userId, User user);
 
-    User getUserById(int userId);
+    User getUserById(Long userId);
 
     List<User> getAllUsers();
 
-    void deleteUserById(int userId);
+    void deleteUserById(Long userId);
 }
